Add arguments can be marked as optional, add validate upper limit of the number of arguments
Both the min and max number of arguments is validated now. Arguments can
be marked as optional.

By default both the lower and upper limit of the number of arguments is
equal to the number of defined arguments. Optional arguments reduce the
lower limit by one. Multiple arguments remove the upper limit
altogether.

Fixes #5.